docker/azure-mcp/requirements.txt
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-mgmt-datafactory==9.0.0
azure-mgmt-resource==23.0.1
openai==1.54.3
requests==2.31.0